[aml] introduce guisettings overrides
This is useful for packagers as they can override menu defaults in as.xml
rather than patching xbmc code.

Usage example:
  <settingsoverride>
    <override>
      <setting hidden="true" locked="true">input.enablemouse</setting>
      <default>false</default>
    </override>
    <override>
      <setting>lookandfeel.enablerssfeeds</setting>
      <default>false</default>
    </override>
  </settingsoverride>

Hidden: Do not show control in GUI.
Locked: Force a default setting.
Default: Set the default for first-run.

The first would disable the mouse. adding "locked" will force the setting, even
if the user currently has it enabled. In addition, it hides the option from the
settings menu.

The second sets the rss feed option to false by default. The user may enable it
later.

A setting that is locked and not hidden will appear grey'd out. Note that
locking is not always safe. Some settings may depend on a control to be enabled
in order to populate other data. Packager beware.